ConstOptree.xs: first attempt to handle LOGOPs
The goal is to reduce conditions like $^O =~ /linux/.
It turns out to be much more difficult than $^O eq "linux",
because:

1) OP_MATCH is always created empty, and only then it gets
"bound" to its arguments.  Thus installing PL_check[OP_MATCH]
hook is pointless.  The only hope, it seems, is to fish for
the OP_MATCH later, somewhere up the tree, like in LOGOP.

2) LOGOP cannot be handled via PL_check either, because
in newLOGOP, optimization for constant condition is done
before CHECKOP (and so reducing the condition to a constant
would be too late, LOGOP won't fold).  Note that, unless
the optimization is effective, newLOGOP wraps the
resulting LOGOP into UNOP, either OP_NULL or OP_NOT.

3) Thus the only hope, it seems, is to install OP_NULL
and OP_NOT hooks which will specifically process LOGOPs.
Namely, such a LOGOP has to be destructed (split back)
into its condition (op_first) and its code block (op_other);
the condition should be converted to a constant, and if
that succeeds, new LOGOP must be constructed, in which
the optimization for constant condition will hopefully work.

In this commit, I try to handle LOGOPs like 'require Foo if $]',
without actually dealing with OP_MATCH.  It almost works,
except that the tree that I get is a little bit deficient.
Here's a normal tree:

$ perl -MO=Concise -e 'require Foo if 1'
5  <@> leave[1 ref] vKP/REFC ->(end)
1     <0> enter ->2
2     <;> nextstate(main 1 -e:1) v:{ ->3
4     <1> require sK/1,FOLD ->5
3        <$> const[PV "Foo.pm"] s/BARE ->4

Note that, as per execution order, nextstate points to "Foo.pm".
When folding 'if $]' with the above technique, I get this:

$ perl -Mblib -MO=Concise -MB::ConstOptree -e 'require Foo if $]'
4  <@> leave[1 ref] vKP/REFC ->(end)
1     <0> enter ->2
2     <;> nextstate(main 165 -e:1) v:{ ->3
3     <1> require sK/1,FOLD ->4
-        <$> const[PV "Foo.pm"] s/BARE ->3

In this tree, nextstate points right to require, and its argument,
"Foo.pm", is orphaned.  The result is only too predictable:

$ perl -Mblib -MB::ConstOptree -e 'require Foo if $]'
Missing or undefined argument to require at -e line 1.

So, how do I fold LOGOPs properly?
Can somebody help me?